b/2224837 Changed the plumbing of how alerts/notifications work.
Cut down on the number db queries when process alerts
Fixed a cursor leak
Flash the event name and location in the notification bar
b/2205255 Show the latest event in the notification. Prioritize based on your acceptance response.
b/1544909 Flash green LED for when there's a calendar notifications.
b/2224837 Cleanup alert/notification plumbing in Calendar
b/1735201 Calendar notifications are not updated when locale is changed

Change-Id: I86b6904607b0236fb04719f5782f43674ac6d2bc