b/2378646 Temporary workaround for Marc to use Exchange Calendar.
Changes the check for accounts when starting Calendar to see if any accounts are adding
to the Calendars table instead of looking for a Google account. Should allow Exchange
to work without a google account. This version still bounces users to the Google
account creation wizard if no calendar accounts exist.