b/2313182 Cleaned up launch activity with better multi-account handling.
Launch activity now just checks for the existence of any account, and if not found
bounces users to the account settings activity where they can add a calendar
supported account. If a non-calendar supported account is found it will let users
enter Calendar but they will be unable to create any events.