b/2432256 Rewrite of Calendars to use HashMaps and async updates.
Changed the way changes to sync states are tracked and updated. Now all
changes are tracked in a hash table which only contains differences from
the initial setting. At save these changes are added to an AsyncQueryHandler
update list to be done in the background. Changes to ExpandableListView and
related need to be made to improve UI response during syncs still.