b/2412594 Added an implicit intent filter to the Calendar Launcher.
Added an implicent intent filter to the Calendar launcher so it will no longer be tied
to the package name. And added functionality to allow a data uri to be passed giving a
time to start the calendar at as well as allowing "VIEW" as an extra with "DAY" to go
to the day view directly.