b/2431102 Added check on save to ensure a valid calendar was selected.
EditEvent performs a background query to get the list of calendars available
before setting a default calendar. This made it possible to hit done on an
event before any calendars could be set. Added a check to ensure a valid
calendar has been selected before a save operation is performed on an event.