Some minor code fixes including cursor management.
Two small fixes. First is new events set to start and end on the epoch get moved
to the current time rounded up to the half hour, but if you were on the hour they
were getting bumped to the next hour. Now they stay on the current hour. Second
is some cursor management in SelectCalendars to ensure all cursors get closed
properly.