b/2496655 Changed launcher to use a default detailed view.
prefs for Calendar seems to have lost a detailed view setting. Changed launcher
to use a default one if it isn't found. Also, a bug fix in getPathSegments()
cause an off by one error for one of our checks which caused us to always go to
the current time. Updated timeFromintentInMillis to fix this.

Change-Id: I78bc380590b8bfae6c475f5e302a4e6a5d082258