b/2500471 Pressing back after setting up account will now enter Calendar
This gets us halfway there. When the user finishes setting up an account it sends
them back to the first Add Account screen. This fix makes it so if they press back
at this point they will return to Calendar and be brought to the default view.
Still working on a way to skip returning to the Add Account screen.

Change-Id: Ie203e74286c40eb8ca65818ac4a633443ebae88b