b/2555049 Will only query accounts once when entering Calendars screen
There was noticeable jitter on entering the Calendars screen including
misregistering taps due to the cursor requerying repeatedly during a
sync. This change makes it so we only query on accounts once and then
will requery every five seconds for about a minute. The view will only
be updated if a change has occurred.

Change-Id: I53610836e78d970d452d4c9724a2d3525cd85482