Creates a model for events and moves edit event methods into a helper
This is a step towards separating the UI and the logic for Calendar. This
creates a standard model for representing events programatically. This
model allows all the fields necessary for updating and managing an event
in the db and a few early helper methods for working with them. EditEventHelper
contains most of the logic from the previous EditEvent activity redone to use
the new event model and without reliance on any UI elements and with db access
minimized. It will still need to be modified to use a service for db access.

Change-Id: Ic8f68269b75f43a0f2fd6408196176592d32e699