Creates an EditEventView with as little app logic as possible
This is work on separating the UI and the logic for Calendar. The view
should be responsible only for updating a model in response to user interaction
and notifying the controller when it is finished. All other work should be
handled at the fragment or activity level. This version does not introduce the
use of fragments. They will come in the next version. It does make all the event
queries asynchronous.

Change-Id: If967c77f39bbbe09490a5fe5e78151e828c8e6cf