Switches EditEventHelper to use AsyncQueryService
This change backgrounds the batch save used in EditEventHelper and
the queries in the EditEventActivity. It also creates an abstract
CalendarActivity with a call for lazy creation/access of an AsyncQueryService
object from the activity. Also changes setting duration to prefer using
the start and end to a pre-existing change to fix b/2789706.

Change-Id: If7fe94086f2a5dcfa34dd208358fada370d0bfde