b/2888467 Fixes duplicate attendees
We weren't saving the attendees to the original model and thus it
always looked like all attendees were new. This change copies the
initial attendees list into the original model. Also makes the same
fix for reminder minutes which were not user visible but would cause
reminders to be removed and re-added.

Change-Id: I99da22fd5173793984b3f1fd21ca3642cecffd95