Allow user to return to last view
This overrides onBackPressed and will return the user to
the last view they were in if they have changed views. It
will only remember one previous view and will exit the app
if back is pressed a second time. This will allow a quick
return from month view or edit event.

Change-Id: I810fdc933bed8d0c2d458cd663220c39979b2619