Integrate tz preferences with db. Do not merge
Will now write and read the tz settings from the db correctly.
Utils attempts to keep the preferences in sync with what's in the
db, but currently does not protect against other apps writing to
the db while we're running. This is not currently an issue as no
one else uses these values and they are not public. Also adds Paris
to the list of time zones (b/2992452).

Change-Id: I6156121f0900001d166948a58ca0ef70a91b3314