Make settings screen support two pane mode.
1) Rename CalendarPreferenceActivity to GeneralPreferences, and
   add CalendarSettingsActivity for top-level Activity.

We should retain previous logs for CalendarPreferenceActivity,
and we shouldn't call it "Activity" any more, as it is Fragment.

I'm not sure whether the new "CalendarSettingsActivity" should
be named so, or renaming it to "CalendarPreferenceActivity"
to keep consistency around naming.

2) Add necessary xmls.

Some of them are derived from Settings app, which already has
new two pane structure, so refering its logic seems enough for now.

Change-Id: I6a0b04bb824d06c2842cbd8127e2860c45350987