b/2988876 Stop crash on orientation change
EditEventFragment was triggering a fragment swap in onPause, which
isn't supported. This fixes the behavior of Edit to handle moving
the fragment around correctly. This change makes the behavior save
the event when back is pressed and restore the event if the user
goes home and returns. On orientation change changes get lost,
however. Another change will be made to fix this.

Change-Id: Iaef5eac6668a8e0ee4859f69e5a77ecb9a7a15a3