Fix double fling in DayView and speed up scroll
The double scroll was that during a fling it was actually snapping
back to the current day then getting a goto event that caused a
fling. The speed issue is something with the background image
rendering, so we have temporarily disabled the background image
and a single color background is being drawn instead.

Change-Id: I8a986eb56e93b08051fa6892d41e3a7e4168f39a