Don't update the mini-month until the day or week transition finishes
Previously we would send an event to the controller (to update the
mini-month) as soon as we recognized the fling gesture.  Now, we delay it
until after the ViewSwitcher animation finishes, to avoid a stutter at the
moment you lift your finger.

This also fixes an unreported bug where the mini-month wouldn't get
updated at all if you dragged left or right and then lifted your finger
without moving.  (In that case we'd animate to the next or previous week,
but the mini-month would stay put.)  That's fixed because we now send the
controller event from switchViews() rather than from doFling().

Bug: 3138116
Change-Id: Ib3ea9a9b356d802338dc9b33622a0a8dfcdad28b