Fix broken DayView X position after a failed fling
When handling a fling gesture, we switch to the next or previous DayView
if the fling is long enough, and if it's more horizontal than vertical.

In the case where we *don't* switch, though, we still need to reset
mViewStartX to make sure we don't get left in a partially-scrolled state.

TESTED: Both week and day views.  Originally I could reproduce the bug
very easily, now I can't.

Bug:  3197125
Change-Id: Iafd27d5246d7e0a25890b238ec7d414fc4c7e3b8