b/2781781 Copies attendees to new exception events
The code that saves a response to an event from the popup view
didn't copy the attendees over to the new event exception. This
adds code to do that. A larger change still needs to be done to
make a single code path for getting an event model and saving the
event.

Change-Id: I90f0a6c501f63691048365a58b8646c8934c609b