b/3335099 Adds a home time zone descriptor to edit event
This adds lines for displaying the date and time of an event
that is in a different time zone than the time zone your calendar
is set to.
Also fixed some bugs:
The ro mode was displaying when in the wrong time zone and new
events weren't being created in the home time zone if it was set.

Change-Id: Ieb8a55a0d0e7c9f26f0246a7bf32da25fba53b5e