b/3354202 Fixes delete future events bug
Delete future events was deleting all events in a series because
the code was checking if it was the first event and if so
deleting everything. We were passing in the instanced event instead
of the original event, which caused it to always think the event
was the first in the series. This changes delete to pass in the
original event instead.

Change-Id: I5f3a1a06f1c854bcf3433ebe4ab15301c8c838ed