b/3384018 Makes disabling/enabling sync also update selected
Fixes a bug where turning off all calendars causes the last query
to continue to be displayed in month view and tweaks the sync
setting interaction to be cleaner (so turning sync off on a
calendar also makes it not selected and vice versa). The second
change also acts as a temporary fix for instances not being
cleared when the sync_events state is changed.

Change-Id: I45462d8ff954ef9b249a58040cd055614b655867