b/3157682 Makes the allday area expandable
If there are too many allday events they become impossible to
read. This makes it so that in this case we will instead hide
the excess events with information about how many are hidden
and an easy way to expand the allday area to show all the
allday events. Also fixes b/3472230

Change-Id: I5a73cc20f1edde54d90fc76fb7ac5ebd1260b073