b/3385145 Use current time for new events in monthview
This adds a fix so that if the user taps today and ends up
in the current month (doesn't stop the scroll prematurely)
and then taps the new event button the app will use the
current time instead of the default time for that month.
Also adds a fix so tapping the create event button will
always create an event on the half hour.

Change-Id: I709df6dbab4d216884f755c8703a325ae30cfa91