Another big pass on colors and spacing
This adds color treatment for current and past time and sorts out
colors and spacing for day and week view. It also makes some
tweaks to month view and mini-month view. Mini-month visuals are
still in progress and the current home time needs to find a
better location (and doesn't exist on the phone atm).

Change-Id: I353e9bf63e4ea6ebf094ed544f4b0a9491c1e238