b/5000382, b/4967872, b/3169913 Beat Agenda View into submission
Fixes several bugs around agenda view not going to the right time
or shifting times on rotate. It will now maintain a consistent day
(couldn't get it to focus to a single view of an instance). There
is still one bug where it will move back one spot some of the time
on phones which will then go to the previous day on another
rotation.

Change-Id: Idee591e37b56d6992f38b89a7896c1397de7df0e