Improve linkification of location field
The "location" field of a calendar Event is expected to hold a
location, but sometimes holds a phone number.  The framework's
Linkify code looks at "1600 Amphitheatre Parkway, 94043" and decides
that "94043" is a phone number, which isn't great.  As part of
bug 3328309, linkification of phone numbers was disabled for the
location field.

We now let Linkify look for non-phone-number items, and then use
the fancy libphonenumber routines in a second pass for more reliable
phone number detection.

If we find what looks like a phone number inside a URL, we ignore
it (i.e. we defer to Linkify when we detect an overlapping span).

Bug 4969832

Change-Id: Ibd763ec297c4506aece2e9190d7c6fcf99e5454c