When restoring instance state, make sure we always check if modification mode has been set.
What had been occuring was that the displayEditWhichDialog checked for the presence of mRrule.
Since this is already checked in setModeIfDone before calling to display the dialog,
this check was not necessary and prevented calls to bring up that dialog
when instance state was restored with an unitialized modification.

Fixes b/5388467 Changing orientation loses edit options while editing a recurring event

Change-Id: I9e8a8ef1b24db0098dfdecd89fdfff19f0f8a027