b/5425449 Speed up attendee loading
This makes a change to how attendees being added to a view are
animated so that the entire view will animate in at once. This
dramatically speeds up the time to display the attendees list
but makes it more jumpy for large lists of attendees when it doesn't
get processor time to perform the animation.

Change-Id: I2742c8c1e6b9092bda62439ccca100987f31c0ac