Change tablet month view event layout
Two changes to the way that events are presented in the tablet
month view.  First, we no longer show the times for each event
(which appear on a second line, roughly doubling the vertical space
consumed) unless we can do so for all events on that day.  Second,
all-day events now get a full-width rectangle to make them easily
distinguishable from non-all-day events.

Also, the decision for whether to use framed vs. filled boxes is
made based on whether the attendee status is INVITED, rather than
NONE (the former indicating an invitation you haven't responded to,
the latter indicating you created an event for yourself).  This
brings the month view behavior in sync with the week view.

Bug 5567167

Change-Id: I986b92acc38e70774fa33e2aaeb063878e31c9a3