Forgot to remove old string.
This string got re-formatted to be used from an array in ag/304669
but forgot to remove it from this file as well.
Note that this radio button does not start with any text because
we don't have anything to put there until we know what day of the
week will be repeating until it gets set in updateDialog() of
onCreateView().

Change-Id: I10f25cac33b5ba82d5578c740c43964c757a125d