Fix issue #457 and remove other bounces
WARNING: This commit introduces or greatly amplifies a bug.
The 24th hour on the DayView cannot be selected. This is resolved in a follow up commit.
At this time, there was no known way to circumvent this while still treating this
and the follow up commit as seperate semantical units.
For discussion about the bug see <https://github.com/Etar-Group/Etar-Calendar/issues/693>

This applies the following commits from <https://github.com/rparkins999/Etar-Calendar>:
  36f1e6609eea42a4a8c5ea942983e577dea4a484 Fix issue #457 and remove other bounces
  b043785f79d46e6f24035ad0b6a51c5216bae41d Remove DEBUG_SCROLLING, left in branch DEBUG_SCROLLING
Along with the following changes done by the committer:
  fixup: undo out of scope changes
  fixup: undo non-functional if block reorganizations

Closes <https://github.com/Etar-Group/Etar-Calendar/issues/457>