Don't mess with selected hour when scrolling
This removes DayView.removeSelectedHour(), as it is no longer used.
This fixes the problem with selecting the 24th hour, which was
introduced in the commit "Fix issue #457 and remove other bounces".

This applies the following commits from <https://github.com/rparkins999/Etar-Calendar>:
  dbfed92215570932c7acb4d2d2b4aa7b4e255e3a Fix problem with selecting hour 23
  8f59dc15ed7fc97edcac7d1167696e9aa8dc27b0 Don't mess with selected hour when scrolling
Along with the following change done by the committer:
  fixup: remove now unused removeSelectedHour()

Closes <https://github.com/Etar-Group/Etar-Calendar/issues/693>