Settings: add a button to add remote EteSync accounts.
EteSync is a fully open-source (clients and server), end-to-end encrypted,
and privacy respecting sync for your contacts, calendars and tasks.

This change was kept as close as possible to the existing CalDAV/DAVx5 button
so it made it possible for me to easily add translations for all languages that
already had the CalDAV button translated.