Add recurrence rule support for importing calendar event from .ics file
The recurrence rule (RRULE) property has been added to the VEvent class. This property represents a recurrence rule, such as "FREQ=WEEKLY;COUNT=10;BYDAY=MO". The property list in the VEvent also has been updated to include RRULE.

The ImportActivity class now includes the RRULE property when creating a calendar intent. This allows the recurrence rule of an event to be imported along with other properties.

It creates the correct recurrence rule while importing a recurring calendar event from an .ics file.